requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
intel-openmp==2021.4.0
jax==0.4.29
jaxlib==0.4.29
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.1
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sounddevice==0.4.7
sympy==1.12.1
tbb==2021.12.0
threadpoolctl==3.5.0
torch==2.3.1
typing_extensions==4.12.2
